backend/requirements.txt
fastapi
uvicorn
aiosqlite
python-jose[cryptography]
bcrypt
python-multipart
httpx
